requirements.txt
streamlit
requests
python-docx
pdfminer.six
